Add hll_count

+++ task/src/hll.h
@@ -12,11 +12,11 @@ typedef struct {
     int b;
 } hll;
 
-// size should be a power of 2
+// size should be a power of 2 >= 2
 hll hll_new(size_t size);
 void hll_destroy(hll *sketch);
 void hll_add(hll *sketch, long long id);
-long long hll_count(hll *sketch);
+long double hll_count(hll *sketch);
 void hll_merge(hll *left, hll *right);
 size_t hll_hash(size_t size, long long id);
